+++ computer-repair-service/src/main/java/com/solvd/app/dao/mysql/EmployeeDAO.java
@@ -42,7 +42,7 @@ public class EmployeeDAO extends MySQLDAO implements IEmployeeDAO {
     }
 
     @Override
-    public void removeById(long id) {
+    public void removeById(int id) {
 
     }
 }
